pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "langchain-qwq"
version = "1.0.0"
description = "An integration package connecting Qwen 3, QwQ and LangChain"
authors = [
  "Yiğit Bekir Kaya, PhD <yigit353@gmail.com>",
  "tiebingice <tiebingice123@outlook.com>",
]
readme = "README.md"
repository = "https://github.com/yigit353/langchain-qwq"
license = "MIT"

[tool.poetry.extras]
docs = ["notebook"]
test = [
  "pytest",
  "pytest-asyncio",
  "pytest-socket",
  "pytest-watcher",
  "langchain-tests",
  "python-dotenv",
]
codespell = ["codespell"]
lint = ["ruff"]
typing = ["mypy"]


[tool.mypy]
disallow_untyped_defs = true

[tool.poetry.urls]
"Source Code" = "https://github.com/yigit353/langchain-qwq/tree/main/langchain_qwq"
"Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22qwq%3D%3D0%22&expanded=true"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-core = "^1.0.0"
openai = "^1.70.0"
langchain-openai = "^1.0.0"
python-dotenv = "^1.1.0"
json-repair = "^0.40.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
notebook = "^7.4.2"

[tool.ruff.lint]
select = ["E", "F", "I", "T201"]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5"
markers = [
  "compile: mark placeholder test used to compile integration tests without running them",
]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.codespell]
optional = true

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest-socket = "^0.7.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-socket = "^0.7.0"
pytest-watcher = "^0.3.4"
langchain-tests = "^1.0.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.6"

[tool.poetry.group.lint.dependencies]
ruff = "^0.5"

[tool.poetry.group.typing.dependencies]
mypy = "^1.10"
